requirements.txt
coverage==5.1
hypothesis==5.10.5
matplotlib==3.2.1
moviepy==1.0.2
mypy==0.770
numpy==1.18.4
pandas==1.0.3
pillow-simd==7.0.0.post3
pytest==5.4.1
pytest-cov==2.8.1
scikit-learn==0.22.2
scipy==1.4.1
seaborn==0.10.1
sphinx==3.0.3
sphinx-autoapi==1.3.0
sphinx-autodoc-typehints==1.10.3
sphinx-rtd-theme==0.4.3
statsmodels==0.11.1
